# app/static/swagger.yaml
openapi: 3.0.3
info:
  title: Mechanic Shop API
  version: 1.0.0
  description: API for customers, mechanics, service tickets, and inventory.
servers:
  - url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
    Token:
      type: object
      properties:
        token:
          type: string
...
    Mechanic:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        email:
          type: string
          format: email
        salary:
          type: integer
        password:
          type: string
          writeOnly: true
      required: [name, email, salary, password]
...
paths:
...
  /mechanics/:
    post:
      tags: [Mechanic]
      summary: Create mechanic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mechanic'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mechanic'
        '400':
          description: Validation error
    get:
      tags: [Mechanic]
      summary: List mechanics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mechanic'
  /mechanics/login:
    post:
      tags: [Mechanic]
      summary: Mechanic login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /mechanics/{mechanic_id}:
    put:
      tags: [Mechanic]
      summary: Update mechanic
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: mechanic_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mechanic'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mechanic'
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags: [Mechanic]
      summary: Delete mechanic
      parameters:
        - in: path
          name: mechanic_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Not Found
...